LayoutTests/fast/css/max-height-and-max-width.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head id="myhead">
<script src="../js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner)
    window.testRunner.dumpAsText();

function doTest()
{
    description("This succeeds if max-height and max-width are not set together, but are different values.");

    var box = document.createElement("div");
    box.style['max-height'] = "20px";
    box.style['max-width'] = "30px";

    document.getElementById("body").appendChild(box);

    var style = window.getComputedStyle(box);
    if(style.getPropertyValue("max-height") == style.getPropertyValue("max-width"))
        testFailed("max-width is NOT independent of max-height.");
    else
        testPassed("max-width is independent of max-height.");
}
</script>
</head>
<body onload="doTest()" id="body">
</body>
</html>
